+++ modules/javafx.controls/src/test/java/test/javafx/scene/control/behavior/TextAreaBehaviorTest.java
@@ -51,155 +51,44 @@ public class TextAreaBehaviorTest extends TextInputControlTestBase<TextArea> {
     }
 
     @Test
-    public void testTypingSanity() {
-        execute(
-            "hello",
-            checkText("hello"),
-            BACK_SPACE, BACK_SPACE, "f",
-            checkText("helf")
-        );
-    }
-
-    /** tests event consumption logic */
-    @Test
+    @Override
     public void testConsume() {
-        Assertions.assertTrue(kb.keyPressed(HOME));
-        Assertions.assertFalse(kb.keyReleased(HOME));
-
-        Assertions.assertTrue(kb.keyPressed(A));
-        Assertions.assertTrue(kb.keyTyped(A, "a"));
-        Assertions.assertFalse(kb.keyReleased(A));
-
-        Assertions.assertTrue(kb.keyPressed(DOWN));
-        Assertions.assertFalse(kb.keyReleased(DOWN));
-
-        Assertions.assertFalse(kb.keyPressed(F1));
-        Assertions.assertFalse(kb.keyReleased(F1));
-
-        Assertions.assertFalse(kb.keyPressed(ESCAPE));
-        Assertions.assertFalse(kb.keyReleased(ESCAPE));
-
-//        Assertions.assertFalse(kb.keyPressed(ENTER));
-//        Assertions.assertFalse(kb.keyReleased(ENTER));
-
-        Assertions.assertTrue(kb.keyPressed(SPACE));
-        Assertions.assertTrue(kb.keyTyped(SPACE, " "));
-        Assertions.assertFalse(kb.keyReleased(SPACE));
+        super.testConsume();
     }
 
     @Test
+    @Override
+    public void testConsumeEnter() {
+        // n/a
+    }
+
+    @Test
+    @Override
     public void testCopy() {
-        execute(
-            setText("abcd"),
-            shift(RIGHT), checkSelection(0, 1),
-            shortcut(C), checkSelection(0, 1), checkClipboard("a"),
-            HOME, RIGHT, shift(RIGHT), checkSelection(1, 2),
-            shortcut(INSERT), checkSelection(1, 2), checkClipboard("b"),
-            HOME, RIGHT, RIGHT, shift(RIGHT), checkSelection(2, 3),
-            COPY, checkSelection(2, 3), checkClipboard("c")
-        );
+        super.testCopy();
     }
 
     @Test
+    @Override
     public void testCut() {
-        execute(
-            setText("aa"),
-            shortcut(A), shortcut(X), checkText("", 0), checkClipboard("aa"),
-            setText("bb"),
-            shortcut(A), CUT, shortcut(X), checkText("", 0), checkClipboard("bb")
-        );
-
-        control.setEditable(false);
-        execute(
-            copy("yo"),
-            setText("aa"),
-            shortcut(A), shortcut(X), checkText("aa", 0, 2), checkClipboard("yo"),
-            setText("bb"),
-            shortcut(A), CUT, shortcut(X), checkText("bb", 0, 2), checkClipboard("yo")
-        );
+        super.testCut();
     }
 
     @Test
-    public void testPaste() {
-        execute(
-            setText(null),
-            copy("1"), shortcut(V), checkText("1", 1),
-            copy("2"), PASTE, checkText("12", 2),
-            copy("3"), shift(INSERT), checkText("123", 3)
-        );
-
-        control.setEditable(false);
-        execute(
-            setText(""),
-            copy("1"), shortcut(V), checkText("", 0),
-            copy("2"), PASTE, checkText("", 0),
-            copy("3"), shift(INSERT), checkText("", 0)
-        );
-    }
-
-//    @Test
-//    public void testNavigation() {
-//        execute(
-//            "0123456789", checkSelection(10),
-//            LEFT, LEFT, checkSelection(8),
-//            RIGHT, checkSelection(9),
-//            UP, checkSelection(0),
-//            DOWN, checkSelection(10),
-//            HOME, checkSelection(0),
-//            END, checkSelection(10),
-//            shortcut(HOME), checkSelection(0),
-//            shortcut(END), checkSelection(10)
-//        );
-//    }
-
-//    @Test
-//    public void testDeletion() {
-//        execute(
-//            setText("0123456789"),
-//            END, BACK_SPACE, checkText("012345678"),
-//            shift(BACK_SPACE), checkText("01234567"),
-//            HOME, DELETE, checkText("1234567")
-//        );
-//
-//        control.setEditable(false);
-//        execute(
-//            setText("0123456789"),
-//            END, BACK_SPACE, checkText("0123456789"),
-//            shift(BACK_SPACE), checkText("0123456789"),
-//            HOME, DELETE, checkText("0123456789")
-//        );
-//    }
-
-//    @Test
-//    public void testSelection() {
-//        execute(
-//            setText("abc"),
-//            HOME, shift(RIGHT), checkSelection(0, 1),
-//            END, shift(LEFT), checkSelection(2, 3),
-//            HOME, shift(DOWN), checkSelection(0, 3),
-//            END, checkSelection(3), shift(UP), checkSelection(0, 3),
-//            HOME, checkSelection(0), shift(END), checkSelection(0, 3),
-//            END, checkSelection(3), shift(HOME), checkSelection(0, 3),
-//            HOME, checkSelection(0), shortcut(A), checkSelection(0, 3)
-//        );
-//    }
-
-    @Disabled("JDK-8296266") // FIX
-    @Test
-    public void testRTL() {
-        execute(
-            setText("العربية"),
-            checkSelection(0, 0),
-            RIGHT, checkSelection(1, 1)
-        );
+    @Override
+    public void testNavigation() {
+        // needs graphics
     }
 
     @Test
-    public void testLTR() {
-        execute(
-            setText("abc"),
-            checkSelection(0, 0),
-            RIGHT, checkSelection(1, 1)
-        );
+    @Override
+    public void testDeletion() {
+        // needs graphics
+    }
+
+    @Test
+    @Override
+    public void testSelection() {
+        // needs graphics
     }
 }
